+++ sdk/communication/AzureCommunicationCommon/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 1.3.0 (Unreleased)
+### Features Added
+- Added support for a new communication identifier `TeamsExtensionUserIdentifier` which maps rawIds with format `8:acs:{resourceId}_{tenantId}_{userId}`.
+- Added `isAnonymous` and `assertedId` properties to `PhoneNumberIdentifier`.
+
 ## 1.2.0 (2024-02-23)
 ### Features Added
 - Added support for a new communication identifier `MicrosoftTeamsAppIdentifier`. It will impact any code that previously depended on the use of `UnknownIdentifier` with rawIDs starting with `28:orgid:`, `28:dod:`, or `28:gcch:`.